@@ -1,84 +1,75 @@
 {
-    "name": "SongNumber",
-    "version": "0.0.1",
-    "author": "Florian Chiș",
-    "homepage": "https://github.com/Fl0r14n/",
-    "private": true,
-    "scripts": {
-        "clean": "ionic-app-scripts clean",
-        "build": "ionic-app-scripts build",
-        "lint": "ionic-app-scripts lint",
-        "ionic:build": "ionic-app-scripts build",
-        "ionic:serve": "ionic-app-scripts serve"
+  "name": "SongNumber",
+  "version": "0.0.1",
+  "author": "Florian Chiș",
+  "homepage": "https://github.com/Fl0r14n/",
+  "private": true,
+  "scripts": {
+    "clean": "ionic-app-scripts clean",
+    "build": "ionic-app-scripts build",
+    "lint": "ionic-app-scripts lint",
+    "ionic:build": "ionic-app-scripts build",
+    "ionic:serve": "ionic-app-scripts serve"
+  },
+  "dependencies": {
+    "@angular/common": "5.0.3",
+    "@angular/compiler": "5.0.3",
+    "@angular/compiler-cli": "5.0.3",
+    "@angular/core": "5.0.3",
+    "@angular/forms": "5.0.3",
+    "@angular/http": "5.0.3",
+    "@angular/platform-browser": "5.0.3",
+    "@angular/platform-browser-dynamic": "5.0.3",
+    "@ionic-native/background-mode": "^4.5.2",
+    "@ionic-native/camera": "^4.5.2",
+    "@ionic-native/core": "4.4.0",
+    "@ionic-native/splash-screen": "^4.4.0",
+    "@ionic-native/status-bar": "4.4.0",
+    "@ionic/pro": "1.0.16",
+    "@ionic/storage": "^2.1.3",
+    "@ngx-translate/core": "^9.0.2",
+    "@ngx-translate/http-loader": "^2.0.1",
+    "acidhax.cordova.chromecast": "https://github.com/Fl0r14n/cordova-chromecast",
+    "cordova-android": "7.0.0",
+    "cordova-browser": "5.0.2",
+    "cordova-chromecast": "git+https://github.com/Fl0r14n/cordova-chromecast.git",
+    "cordova-plugin-background-mode": "^0.7.2",
+    "cordova-plugin-camera": "^3.0.0",
+    "cordova-plugin-device": "1.1.4",
+    "cordova-plugin-ionic-webview": "^1.1.16",
+    "cordova-plugin-splashscreen": "^4.0.3",
+    "cordova-plugin-whitelist": "^1.3.1",
+    "ionic-angular": "3.9.2",
+    "ionic-plugin-keyboard": "^2.2.1",
+    "ionicons": "3.0.0",
+    "rxjs": "5.5.2",
+    "sw-toolbox": "3.6.0",
+    "ubuntu-fontface": "^0.1.13",
+    "zone.js": "0.8.18"
+  },
+  "devDependencies": {
+    "@ionic/app-scripts": "3.1.6",
+    "typescript": "2.4.2"
+  },
+  "description": "SongNumber: a small mobile app that will allow user to set a song number from a list of song books and cast it",
+  "cordova": {
+    "plugins": {
+      "cordova-plugin-camera": {},
+      "cordova-plugin-splashscreen": {},
+      "acidhax.cordova.chromecast": {},
+      "cordova-plugin-background-mode": {},
+      "ionic-plugin-keyboard": {},
+      "cordova-plugin-whitelist": {},
+      "cordova-plugin-device": {},
+      "cordova-plugin-ionic-webview": {}
     },
-    "dependencies": {
-        "@angular/common": "4.1.3",
-        "@angular/compiler": "4.1.3",
-        "@angular/compiler-cli": "4.1.3",
-        "@angular/core": "4.1.3",
-        "@angular/forms": "4.1.3",
-        "@angular/http": "4.1.3",
-        "@angular/platform-browser": "4.1.3",
-        "@angular/platform-browser-dynamic": "4.1.3",
-        "@ionic-native/background-mode": "^4.2.1",
-        "@ionic-native/camera": "^4.2.1",
-        "@ionic-native/core": "3.12.1",
-        "@ionic-native/splash-screen": "^3.12.1",
-        "@ionic-native/status-bar": "3.12.1",
-        "@ionic/storage": "^2.0.1",
-        "@ngx-translate/core": "^8.0.0",
-        "@ngx-translate/http-loader": "^0.1.0",
-        "cordova-android": "^6.2.3",
-        "cordova-browser": "^4.1.0",
-        "cordova-chromecast": "git+https://github.com/Fl0r14n/cordova-chromecast.git",
-        "cordova-plugin-add-swift-support": "^1.7.0",
-        "cordova-plugin-background-mode": "^0.7.2",
-        "cordova-plugin-camera": "^2.4.1",
-        "cordova-plugin-compat": "^1.2.0",
-        "cordova-plugin-device": "^1.1.4",
-        "cordova-plugin-ionic": "^1.1.9",
-        "cordova-plugin-splashscreen": "^4.0.3",
-        "cordova-plugin-statusbar": "^2.2.2",
-        "cordova-plugin-whitelist": "^1.3.1",
-        "ionic-angular": "3.6.1",
-        "ionic-plugin-keyboard": "^2.2.1",
-        "ionicons": "3.0.0",
-        "rxjs": "5.4.0",
-        "sw-toolbox": "3.6.0",
-        "ubuntu-fontface": "^0.1.13",
-        "zone.js": "0.8.12",
-        "acidhax.cordova.chromecast": "https://github.com/Fl0r14n/cordova-chromecast"
-    },
-    "devDependencies": {
-        "@ionic/app-scripts": "^2.1.4",
-        "typescript": "2.3.4"
-    },
-    "description": "SongNumber: a small mobile app that will allow user to set a song number from a list of song books and cast it",
-    "cordova": {
-        "plugins": {
-            "cordova-plugin-device": {},
-            "cordova-plugin-splashscreen": {},
-            "cordova-plugin-statusbar": {},
-            "cordova-plugin-whitelist": {},
-            "ionic-plugin-keyboard": {},
-            "cordova-plugin-camera": {},
-            "cordova-plugin-background-mode": {},
-            "cordova-plugin-ionic": {
-                "APP_ID": "20adb2ff",
-                "CHANNEL_NAME": "Master",
-                "UPDATE_METHOD": "auto",
-                "UPDATE_API": "https://api.ionicjs.com",
-                "MAX_STORE": "2"
-            },
-            "acidhax.cordova.chromecast": {}
-        },
-        "platforms": [
-            "android",
-            "browser"
-        ]
-    },
-    "config": {
-        "ionic_copy": "./config/copy.config.js",
-        "ionic_sass": "./config/sass.config.js"
-    }
+    "platforms": [
+      "browser",
+      "android"
+    ]
+  },
+  "config": {
+    "ionic_copy": "./config/copy.config.js",
+    "ionic_sass": "./config/sass.config.js"
+  }
 }
